+++ framework/widgets/ActiveField.php
@@ -145,6 +145,10 @@ class ActiveField extends Component
      * it is maintained by various methods of this class.
      */
     public $parts = [];
+    /**
+     * @var bool adds ARIA HTML attributes `aria-required` and `aria-invalid` for inputs
+     */
+    public $addAriaAttributes = false;
 
     /**
      * @var string this property holds a custom input id if it was set using [[inputOptions]] or in one of the
@@ -156,6 +160,11 @@ class ActiveField extends Component
      * @var bool if "for" field label attribute should be skipped.
      */
     private $_skipLabelFor = false;
+
+    /**
+     * @var bool whether `aria-invalid` HTML attribute should be toggled by validation on client
+     */
+    private $_skipClientAriaInvalid = true;
 
 
     /**
@@ -357,6 +366,7 @@ class ActiveField extends Component
     public function input($type, $options = [])
     {
         $options = array_merge($this->inputOptions, $options);
+        $this->addAriaAttributes($options);
         $this->adjustLabelFor($options);
         $this->parts['{input}'] = Html::activeInput($type, $this->model, $this->attribute, $options);
 
@@ -383,6 +393,7 @@ class ActiveField extends Component
     public function textInput($options = [])
     {
         $options = array_merge($this->inputOptions, $options);
+        $this->addAriaAttributes($options);
         $this->adjustLabelFor($options);
         $this->parts['{input}'] = Html::activeTextInput($this->model, $this->attribute, $options);
 
@@ -428,6 +439,7 @@ class ActiveField extends Component
     public function passwordInput($options = [])
     {
         $options = array_merge($this->inputOptions, $options);
+        $this->addAriaAttributes($options);
         $this->adjustLabelFor($options);
         $this->parts['{input}'] = Html::activePasswordInput($this->model, $this->attribute, $options);
 
@@ -455,6 +467,7 @@ class ActiveField extends Component
         if (!isset($this->form->options['enctype'])) {
             $this->form->options['enctype'] = 'multipart/form-data';
         }
+        $this->addAriaAttributes($options);
         $this->adjustLabelFor($options);
         $this->parts['{input}'] = Html::activeFileInput($this->model, $this->attribute, $options);
 
@@ -474,6 +487,7 @@ class ActiveField extends Component
     public function textarea($options = [])
     {
         $options = array_merge($this->inputOptions, $options);
+        $this->addAriaAttributes($options);
         $this->adjustLabelFor($options);
         $this->parts['{input}'] = Html::activeTextarea($this->model, $this->attribute, $options);
 
@@ -521,6 +535,7 @@ class ActiveField extends Component
             $options['label'] = null;
             $this->parts['{input}'] = Html::activeRadio($this->model, $this->attribute, $options);
         }
+        $this->addAriaAttributes($options);
         $this->adjustLabelFor($options);
 
         return $this;
@@ -567,6 +582,7 @@ class ActiveField extends Component
             $options['label'] = null;
             $this->parts['{input}'] = Html::activeCheckbox($this->model, $this->attribute, $options);
         }
+        $this->addAriaAttributes($options);
         $this->adjustLabelFor($options);
 
         return $this;
@@ -594,6 +610,7 @@ class ActiveField extends Component
     public function dropDownList($items, $options = [])
     {
         $options = array_merge($this->inputOptions, $options);
+        $this->addAriaAttributes($options);
         $this->adjustLabelFor($options);
         $this->parts['{input}'] = Html::activeDropDownList($this->model, $this->attribute, $items, $options);
 
@@ -622,6 +639,7 @@ class ActiveField extends Component
     public function listBox($items, $options = [])
     {
         $options = array_merge($this->inputOptions, $options);
+        $this->addAriaAttributes($options);
         $this->adjustLabelFor($options);
         $this->parts['{input}'] = Html::activeListBox($this->model, $this->attribute, $items, $options);
 
@@ -641,6 +659,7 @@ class ActiveField extends Component
      */
     public function checkboxList($items, $options = [])
     {
+        $this->addAriaAttributes($options);
         $this->adjustLabelFor($options);
         $this->_skipLabelFor = true;
         $this->parts['{input}'] = Html::activeCheckboxList($this->model, $this->attribute, $items, $options);
@@ -660,6 +679,7 @@ class ActiveField extends Component
      */
     public function radioList($items, $options = [])
     {
+        $this->addAriaAttributes($options);
         $this->adjustLabelFor($options);
         $this->_skipLabelFor = true;
         $this->parts['{input}'] = Html::activeRadioList($this->model, $this->attribute, $items, $options);
@@ -698,6 +718,7 @@ class ActiveField extends Component
         $config['attribute'] = $this->attribute;
         $config['view'] = $this->form->getView();
         if (isset($config['options']) && isset(class_parents($class)['yii\widgets\InputWidget'])) {
+            $this->addAriaAttributes($config['options']);
             $this->adjustLabelFor($config['options']);
         }
         $this->parts['{input}'] = $class::widget($config);
@@ -731,8 +752,8 @@ class ActiveField extends Component
             return [];
         }
 
-        $enableClientValidation = $this->enableClientValidation || $this->enableClientValidation === null && $this->form->enableClientValidation;
-        $enableAjaxValidation = $this->enableAjaxValidation || $this->enableAjaxValidation === null && $this->form->enableAjaxValidation;
+        $enableClientValidation = $this->getEnableClientValidation();
+        $enableAjaxValidation = $this->getEnableAjaxValidation();
 
         if ($enableClientValidation) {
             $validators = [];
@@ -780,6 +801,10 @@ class ActiveField extends Component
             $options['validate'] = new JsExpression("function (attribute, value, messages, deferred, \$form) {" . implode('', $validators) . '}');
         }
 
+        if ($this->addAriaAttributes && !$this->_skipClientAriaInvalid) {
+            $options['ariaInvalidToggle'] = true;
+        }
+
         // only get the options that are different from the default ones (set in yii.activeForm.js)
         return array_diff_assoc($options, [
             'validateOnChange' => true,
@@ -788,7 +813,27 @@ class ActiveField extends Component
             'validationDelay' => 500,
             'encodeError' => true,
             'error' => '.help-block',
+            'ariaInvalidToggle' => false,
         ]);
+    }
+
+    /**
+     * Checks if client validation enabled for the field
+     * @return bool
+     */
+    protected function getEnableClientValidation()
+    {
+        return $this->enableClientValidation
+            || $this->enableClientValidation === null && $this->form->enableClientValidation;
+    }
+
+    /**
+     * Checks if ajax validation enabled for the field
+     * @return bool
+     */
+    protected function getEnableAjaxValidation()
+    {
+        return $this->enableAjaxValidation || $this->enableAjaxValidation === null && $this->form->enableAjaxValidation;
     }
 
     /**
@@ -799,5 +844,22 @@ class ActiveField extends Component
     protected function getInputId()
     {
         return $this->_inputId ?: Html::getInputId($this->model, $this->attribute);
+    }
+
+    /**
+     * Adds ARIA attributes to the input options
+     * @param $options array input options
+     */
+    protected function addAriaAttributes(&$options)
+    {
+        if ($this->addAriaAttributes) {
+            if (!isset($options['aria-required'])) {
+                $options['aria-required'] = $this->model->isAttributeRequired($this->attribute) ? 'true' : 'false';
+            }
+            if (!isset($options['aria-invalid'])) {
+                $options['aria-invalid'] = $this->model->hasErrors($this->attribute) ? 'true' : 'false';
+                $this->_skipClientAriaInvalid = false;
+            }
+        }
     }
 }